service/requirements.txt
google-api-python-client
google-cloud-storage
GoogleAppEngineCloudStorageClient
WTForms